+++ .dev.vars.example
@@ -5,4 +5,3 @@ AZURE_GATEWAY_URL=
 OPENAI_API_KEY=
 OPENAI_GATEWAY_URL=
 GROQ_CLOUD_TOKEN=
-GLOBALGPT_API_KEY=
