+++ src/proxy/azure.ts
@@ -1,39 +1,5 @@
-import { models, generativeModelMappings, isNotEmpty } from '../utils';
-
-// support printer mode and add newline
-async function stream(readable: ReadableStream, writable: WritableStream) {
-  const reader = readable.getReader();
-  const writer = writable.getWriter();
-
-  const encoder = new TextEncoder();
-  const decoder = new TextDecoder();
-
-  const newline = '\n';
-  const delimiter = '\n\n';
-  const encodedNewline = encoder.encode(newline);
-
-  let buffer = '';
-  while (true) {
-    let { value, done } = await reader.read();
-    if (done) {
-      break;
-    }
-    buffer += decoder.decode(value, { stream: true });
-    let lines = buffer.split(delimiter);
-
-    for (let i = 0; i < lines.length - 1; i++) {
-      await writer.write(encoder.encode(lines[i] + delimiter));
-    }
-
-    buffer = lines[lines.length - 1];
-  }
-
-  if (buffer) {
-    await writer.write(encoder.encode(buffer));
-  }
-  await writer.write(encodedNewline);
-  await writer.close();
-}
+import { models, generativeModelMappings, isNotEmpty, requestFactory } from '../utils';
+import { streamByOpenAI } from '../utils/stream';
 
 /**
  * Due to varying quota limits across different regions when deploying models on Microsoft Azure,
@@ -56,10 +22,7 @@ const getResourceNameAndToken = (model: string, env: Env) => {
   return resourceInfo;
 };
 
-const proxy: IProxy = async (request: Request, body: any, url: URL, env: Env) => {
-  // Remove the leading /v1/ from url.pathname
-  const action = url.pathname.replace(/^\/+v1\/+/, '');
-
+const proxy: IProxy = async (action: string, body: any, env: Env) => {
   let models_maping = Object.create(null);
   let [gpt35, gpt4] = ['', ''];
   if (isNotEmpty(env.AZURE_DEPLOY_NAME)) {
@@ -75,7 +38,7 @@ const proxy: IProxy = async (request: Request, body: any, url: URL, env: Env) =>
   if (action === 'models') return models(models_maping);
 
   const payload = {
-    method: request.method,
+    method: 'POST',
     headers: {
       'Content-Type': 'application/json',
       'api-key': '',
@@ -101,14 +64,13 @@ const proxy: IProxy = async (request: Request, body: any, url: URL, env: Env) =>
     ? `${env.AZURE_GATEWAY_URL}/${resourceName}/${deployName}/${action}?api-version=${API_VERSION}`
     : `https://${resourceName}.openai.azure.com/openai/deployments/${deployName}/${action}?api-version=${API_VERSION}`;
 
-  let response = await fetch(fetchAPI, payload);
-  response = new Response(response.body, response);
-  response.headers.set('Access-Control-Allow-Origin', '*');
+  const requestFunc = requestFactory(fetchAPI);
+  const response = await requestFunc(payload);
 
-  if (body?.stream != true) return response;
+  if (!body?.stream) return response;
 
-  let { readable, writable } = new TransformStream();
-  stream(response.body as ReadableStream, writable);
+  const { readable, writable } = new TransformStream();
+  streamByOpenAI(response.body as ReadableStream, writable, env, payload, requestFunc);
   return new Response(readable, response);
 };
